% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveFirstRowCol.R.R
\name{moveFirstRowCol}
\alias{moveFirstRowCol}
\title{Move First Row to Column Names and First Column to Row Names}
\usage{
moveFirstRowCol(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with updated column names and row names.
}
\description{
This function moves the first row of a data frame to the column names and the first column to the row names.
}
\examples{
df <- data.frame(V1 = c("ID", "A", "B"), V2 = c("Gene1", 1, 2), V3 = c("Gene2", 3, 4))
moveFirstRowCol(df)
}
